+++ CHANGELOG.md
@@ -18,6 +18,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Don't accept requests signed with DIDs.
 
+### Fixed
+
+- Don't parse `#1` as a hashtag.
+
 ## [4.4.0] - 2025-06-05
 
 ### Added